update compile script

+++ compile.sh
@@ -22,5 +22,9 @@ cp books/build/libs/books-0.0.1-SNAPSHOT.jar artifacts/books-app.jar
 ./gradlew clean :books:nativeCompile
 cp books/build/native/nativeCompile/books artifacts/books-app
 
-./gradlew -Pbp=optimized clean :books:nativeCompile
-cp books/build/native/nativeCompile/books artifacts/books-app
+if [[ -f default.iprof ]]; then
+  ./gradlew -Pbp=optimized clean :books:nativeCompile
+  cp books/build/native/nativeCompile/books artifacts/books-app
+else
+  echo "No profile data, please generate profile data before compile (see README.md)"
+fi
